backend/requirements.txt
fastapi==0.133.0
uvicorn==0.41.0
anthropic==0.83.0
pandas==3.0.1
faiss-cpu==1.13.2
sentence-transformers==5.2.3
rank-bm25==0.2.2
python-dotenv==1.2.1
numpy==2.4.2
